test: refactor whenever type tests into dedicated files
- Split existing tests into `instant_test.py` and `zoned_datetime_test.py` to improve maintainability.
- Extract the `WheneverModel` definition into a shared `models.py` file to reduce duplication across test suites.

Generated-by: aiautocommit
